requirements.txt
aioconsole
beautifulsoup4
discord.py
osuapi
python-dateutil
requests
